Add CLI tests
